pyproject.toml
[project]
name = "rag-techiques"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bs4>=0.0.2",
    "langchain>=1.0.5",
    "langchain-cohere>=0.5.0",
    "langchain-qdrant>=1.1.0",
    "openai>=2.7.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
]
